LLM/realtime-summary/requirements.txt
numpy
sounddevice
webrtcvad
faster-whisper
llama-cpp-python
python-docx  # optional, only if you want DOCX export
